chore: rename `BenefitCtaLinkLight` into `BenefitCtaLink`
